Fixed shuffle code. Added Dealing of hands
Fixed the shuffle code. Mistakenly thought the gameobjects I use to represent the array was the actual array, so i ended up using code that changes the values in the array but not the position of the array. 

Added a variable to safe guard the order of array and rest the array for every button press. 

Added visual representations to see the cards value